refactor ActionMenu using Popover
